Correctly calculate `consumedSystemFlux` for consumers
This commit makes two different changes:
- It rewrites `HeatLoop.AllocateFlux` to avoid the issue where
  `consumedSystemFlux` is set to 0 if a system's `totalSystemFlux` is
  equal to `totalFlux` when it is checked.
- It changes the total flux available to allocate to also take into
  account negative flux used to lower the temperature of the loop.

With these two changes `consumedSystemFlux` should now be accurate when
the loop temperature is >= its nominal temperature. It should now be
possible to actually use it to drive downstream things based on the
amount of heat used (e.g. consuming resources, transferring heat, or
powering an engine).

I have taken care to make sure that the behaviour stays consistent when
below the nominal temperature: flux will still be allocated to consumers
based on the positive flux, even if this has no real effect on the loop
itself.

Fixes #138